refactor: use f64 throughout and improve thread spawning pattern
- Use f64 for all audio/visual data (AudioSamples, AudioState, ControlVoltage)
- Convert from f32 at cpal boundary only
- Introduce thread_body() pattern returning impl FnOnce() for clearer ownership
- Apply to both DspEngine and Inference
- Add #[must_use] to App::new
- Fix clippy pedantic warnings

+++ src/audio.rs
@@ -13,4 +13,4 @@ pub use error::AudioDriverError;
 pub const HOP_SIZE: usize = 512;
 
 /// Audio samples buffer sent from `AudioDriver` to `DspEngine`.
-pub type AudioSamples = [f32; HOP_SIZE];
+pub type AudioSamples = [f64; HOP_SIZE];
